requirements.txt
python-docx 
requests 
tiktoken 
openai 
pillow